+++ 5.0/BlazorSample_Server/Services/TransitiveTransientDisposableDependency.cs
@@ -1,4 +1,4 @@
-namespace BlazorSample;
+namespace BlazorSample.Services;
 
 public class TransitiveTransientDisposableDependency
     : ITransitiveTransientDisposableDependency, IDisposable
