+++ 6.0/BlazorSample_Server/Program.cs
@@ -1,3 +1,5 @@
+using BlazorSample.Services;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
